# [NPU] change Add to AddN in sum npu op