<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization

### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs

### Describe
<!-- Describe what this PR does -->

use `AddN` instead of `Add`.

- before (use Add)

<img width="832" alt="图片" src="https://user-images.githubusercontent.com/26408901/117626895-31068680-b1aa-11eb-993b-81005f625e8f.png">

As shown in the timeline, the time consumption of `ClipGrad` is 144ms (439ms - 583ms), and the time consumption of `Add` in `ClipGrad` is 23ms (521ms - 544ms), which accounts for 16.0%. 

- after (use AddN)
<img width="741" alt="图片" src="https://user-images.githubusercontent.com/26408901/117627561-df123080-b1aa-11eb-80ae-02553172fbd8.png">

As shown in the timeline, the time consumption of `ClipGrad` is 138ms (471ms - 609ms), and the time consumption of `AddN` in `ClipGrad` is 3ms (562ms - 565ms), which accounts for 2.2%. 

### Performance

Speed up: xxx tokens/s -> xxx tokens/s, +xxx  %